requirements.txt
openai
python-dotenv
langchain_openai
langchain_ollama
langchain-ollama
langchain-google-genai
langchain_core
langchain-aws